Update blobs.py contracts
Contract changes:
- upload: does not require anymore to have a file to upload. It takes the name of the entity to upload and the data.
- download: it returns directly the data requested

If a pickle dump or load operation is needed, is now up to an ObjectPersistenceManager subclass to take care of that operation.

The advantage of this update is that client classes now have to pass the data to the upload method, and receive them from the download method. The logic is simplified, and clients have less repsonsibility.

Also, with this update, the ObjecPersistenceManager is now ready to support databases as persistence tools.